Move the types to api.ts
