+++ ui/src/utils/api.ts
@@ -1,3 +1,17 @@
+export interface FieldData {
+  value?: string;
+  confidence?: string;
+}
+
+export interface ExtractedData {
+  [key: string]: FieldData;
+}
+
+export interface VerifiedData {
+  extracted_data?: ExtractedData;
+  [key: string]: unknown;
+}
+
 export async function authorizedFetch(
   url: string,
   options: RequestInit = {}
