--- ui/src/utils/types.ts
@@ -1,13 +1,0 @@
-export interface FieldData {
-  value?: string;
-  confidence?: string;
-}
-
-export interface ExtractedData {
-  [key: string]: FieldData;
-}
-
-export interface VerifiedData {
-  extracted_data?: ExtractedData;
-  [key: string]: any;
-}
